+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Remove support for adding constraints with unknown variables. It was unused and made certain assumptions hard to enforce.
 - (Internal) The `initial_vars` object in config is now an array
+- Drop support for dynamically get the list of vars to target
 
 ## v2.1.2: